pyproject.toml
[project]
name = "projview"
version = "0.1.0"
description = "Store projects folders flat and construct various views"
authors = [
    {name = "Ethan Dawes", email = "ethand.python@gmail.com"},
]
dependencies = [
    "winshell>=0.6",
    "inquirerpy>=0.3.4",
    "pywin32>=310",
    "pyyaml>=6.0.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
projects = "project_cli:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.pdm]
distribution = true

